requirements.txt
Django
django-ninja
pydantic
Jinja2
mysqlclient
django_redis
django-environ
django_static_collector
PyJWT
python-json-logger
arrow
click
uwsgi
